T10Q1 Child Class Bulk Discount int minimum @ 231219, 16:38

+++ src/Tutorial/T10/Q1BulkDiscount.java
@@ -7,7 +7,7 @@ public class Q1BulkDiscount extends Q1DiscountPolicy{
     
     private double minimum, discountRate;
     
-    public Q1BulkDiscount(int minimum, double discountRate) {
+    public Q1BulkDiscount(double minimum, double discountRate) {
         this.minimum = minimum;
         this.discountRate = discountRate;
     }
